Former-commit-id: 152faf0238fa6840f49832ece8f8455a06634789
Former-commit-id: c2749cc503d094528d9b8aca7124f8016f8fc410

+++ server/src/main/java/org/icgc/dcc/validation/cascading/StructuralCheckFunction.java
@@ -87,7 +87,7 @@ public class StructuralCheckFunction extends BaseOperation implements Function {
 
     TupleEntry arguments = functionCall.getArguments();
 
-    int offset = functionCall.getArguments().getInteger(ValidationFields.OFFSET_FIELD_NAME);
+    long offset = functionCall.getArguments().getLong(ValidationFields.OFFSET_FIELD_NAME);
     TupleState tupleState = new TupleState(offset);
 
     String line = arguments.getString(StructuralCheckFunction.LINE_FIELD_NAME);
@@ -120,7 +120,7 @@ public class StructuralCheckFunction extends BaseOperation implements Function {
       tupleState.reportError(ValidationErrorCode.STRUCTURALLY_INVALID_ROW_ERROR,//
           ROW_LEVEL_ERROR_COLUMN_NAME, // because we don't have a specific column to report this on (TODO: revisit not
                                        // elegant)
-          dataSize, headerSize);
+          dataSize, headerSize); // we use dataSize as the value, and headerSize as an expected value parameter
     }
     return adjustedValues;
   }
